Avoid narrowing in array initializator
Gcc 4.7 warns that this kind of narrowing is illegal in C++11, so
we can just as well fix that right now.

In GuiFontMetrics this is fixed by chaning the types in the AscendDescend struct
from short int to int.
In ParagraphMetrics it is fixed by addint explicit casts.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@40586 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/qt4/GuiFontMetrics.h
@@ -66,8 +66,8 @@ private:
 	mutable QHash<char_type, int> width_cache_;
 
 	struct AscendDescend {
-		short int ascent;
-		short int descent;
+		int ascent;
+		int descent;
 	};
 	/// Cache of char ascends and descends
 	mutable QHash<char_type, AscendDescend> metrics_cache_;
